Cleaned code
