+++ src/main/java/com/app/screen/controller/MainController.java
@@ -2,6 +2,7 @@ package com.app.screen.controller;
 
 import com.app.service.AppMain;
 import com.app.service.graph.Graph;
+import com.app.service.graph.GraphService;
 import com.app.service.notification.NotificationType;
 import javafx.application.Platform;
 import javafx.embed.swing.SwingNode;
@@ -129,86 +130,13 @@ public class MainController implements Initializable {
 
     public void runMeasurement(MouseEvent event) {
         // TODO: run measurement and graph
-//        VBox1 = new VBox();
-//        JFrame frame=new JFrame("Chart");
-        Graph rtcp=new Graph("Chart", "Resistance", "Capacity", "Frequency");
-//        ChartViewer chartViewer = new ChartViewer(rtcp.chart);
-//        chartViewer.setPrefHeight(680);
-//        chartViewer.setPrefWidth(260);
-//        chartViewer.setVisible(true);
 
-//        lowerPane = new Pane();
-//        lowerPane.setPrefSize(680, 260);
-//        ChartCanvas canvas = chartViewer.getCanvas();
-//        canvas.heightProperty().bind(lowerPane.heightProperty());
-//        canvas.widthProperty().bind(lowerPane.widthProperty());
-//        lowerPane.getChildren().addAll(canvas);
-//
-        ChartPanel chartPanel = new ChartPanel(rtcp.chart);
-        chartPanel.setPreferredSize(new java.awt.Dimension(680, 260));
-//
-//        upperPane = new AnchorPane();
-//        upperPane.setPrefSize(680, 260);
-//        upperPane.getChildren().addAll(chartViewer);
-//        upperPane.setVisible(true);
-//        VBox1.getChildren().add(upperPane);
-//        Scene primaryScene = new Scene(VBox1);
-
-
-        final SwingNode swingNode = new SwingNode();
-        swingNode.setContent(chartPanel);
-        upperPane.getChildren().addAll(swingNode);
-//
-//        Stage stage = new Stage();
-//        stage.setScene(new Scene(upperPane, 0, 0));
-//        stage.show();
-
-
-//        frame.getContentPane().add(rtcp,new BorderLayout().CENTER);
-//        frame.pack();
-//        frame.setVisible(true);
-
-//        (new Thread(rtcp)).start(); 1st version
-
-
-
-//        TimerTask task = new TimerTask() {
-//            public void run() {
-//                System.out.println("Task performed on: " + new Date() + "n" + "Thread's name: " + Thread.currentThread().getName()); }
-//                };
-//        Timer timer = new Timer("Timer");
-//        long delay = 1000L;
-//        timer.schedule(task, delay);
-
-        // real-time plotting
-        final double[] poc = {0};
-        //now make your timer
-        int delay = 500; //milliseconds
         try {
-
-
-            ActionListener timerAction = new ActionListener() {
-                public void actionPerformed(ActionEvent evt) {
-                    //add new data point and actualize graph
-                    rtcp.series1.add(poc[0], (double) (Math.random() * 20 + 80));  // tu budu pribudat values
-                    rtcp.series2.add(poc[0], (double) (Math.random() * 20 + 80));  // tu budu pribudat values
-                    poc[0]++;
-
-                }
-            };
-            new Timer(delay, timerAction).start();
-        } catch (NoSuchMethodError e) {
+            GraphService graphService = new GraphService(upperPane);
+            graphService.createGraphRun();
+        } catch (NoSuchMethodError e) { // catches exception/error that occurs always when running graph, but works
+            throw e;
         }
-
-//        frame.addWindowListener(new WindowAdapter()
-//        {
-//            public void windowClosing(WindowEvent windowevent)
-//            {
-//                System.exit(0);
-//            }
-//
-//        });
-
     }
 
     public void toggleAutoSave(MouseEvent event) {
